moved centralized request authorization logic to more definitive type

+++ backend/app/Http/Controllers/Api/AttachmentsController.php
@@ -2,25 +2,20 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Requests\Tasks\StoreTaskRequest;
-use App\Http\Requests\Tasks\UpdateTaskRequest;
+use App\Http\Requests\Tasks\StoreAttachmentRequest;
+use App\Http\Requests\Tasks\UpdateAttachmentRequest;
 use App\Models\Attachment;
-use App\Models\Comment;
 use App\Models\Task;
 use Illuminate\Http\JsonResponse;
 
 class AttachmentsController extends ApiController
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Attachment::class);
-    }
-
     /**
      * Display a listing of the resource.
      */
     public function index(Task $task): JsonResponse
     {
+        $this->authorize('viewAny', [Attachment::class]);
         $comments = $task->comments()->get();
         return $this->jsonResponse(compact('comments'));
     }
@@ -29,10 +24,15 @@ class AttachmentsController extends ApiController
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreTaskRequest $request, Task $task): JsonResponse
+    public function store(StoreAttachmentRequest $request, Task $task): JsonResponse
     {
-        $comment = Comment::create($request->validated());
-        return $this->jsonResponse($comment, 201);
+        $this->authorize('create', [Attachment::class]);
+
+        $attachment = $task
+            ->attachments()
+            ->create($request->validated());
+
+        return $this->jsonResponse($attachment, 201);
     }
 
     /**
@@ -40,14 +40,18 @@ class AttachmentsController extends ApiController
      */
     public function show(Task $task, Attachment $attachment): JsonResponse
     {
+        $this->authorize('view', [Attachment::class, $task, $attachment]);
+
         return $this->jsonResponse($attachment);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateTaskRequest $request, Task $task, Attachment $attachment): JsonResponse
+    public function update(UpdateAttachmentRequest $request, Task $task, Attachment $attachment): JsonResponse
     {
+        $this->authorize('view', [Attachment::class, $task, $attachment]);
+
         $attachment->update($request->validated());
         return $this->jsonResponse($attachment);
     }
@@ -57,6 +61,7 @@ class AttachmentsController extends ApiController
      */
     public function destroy(Task $task, Attachment $attachment): JsonResponse
     {
+        $this->authorize('delete', [Attachment::class, $task, $attachment]);
         $attachment->delete();
         return $this->jsonResponse($attachment);
     }
