+++ backend/app/Http/Controllers/Api/TasksController.php
@@ -13,16 +13,13 @@ use Illuminate\Http\JsonResponse;
 
 class TasksController extends ApiController
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Task::class);
-    }
-
     /**
      * Display a listing of the resource.
      */
     public function index(Project $project, Board $board): JsonResponse
     {
+        $this->authorize('viewAny', [Task::class, $project, $board]);
+
         $tasks = $board->tasks()->get();
         return $this->jsonResponse(compact('tasks'));
     }
@@ -33,6 +30,8 @@ class TasksController extends ApiController
      */
     public function store(StoreTaskRequest $request, Project $project, Board $board): JsonResponse
     {
+        $this->authorize('create', [Task::class, $project, $board]);
+
         /** @var Task $task */
         $task = $board->tasks()->create($request->except(['assignees', 'tags']));
 
@@ -55,6 +54,8 @@ class TasksController extends ApiController
      */
     public function show(Project $project, Board $board, Task $task): JsonResponse
     {
+        $this->authorize('view', [Task::class, $project, $board, $task]);
+
         $task = $task->toArray();
         $task['assignees'] = TaskUser::where('task_id', $task['id'])->pluck('user_id')->toArray();
         $task['tags'] = TaskTag::where('task_id', $task['id'])->pluck('tag_id')->toArray();
@@ -67,6 +68,8 @@ class TasksController extends ApiController
      */
     public function update(UpdateTaskRequest $request, Project $project, Board $board, Task $task): JsonResponse
     {
+        $this->authorize('update', [Task::class, $project, $board, $task]);
+
         /** @var Task $task */
         $task->update($request->except(['assignees', 'tags']));
 
@@ -93,6 +96,8 @@ class TasksController extends ApiController
      */
     public function destroy(Project $project, Board $board, Task $task): JsonResponse
     {
+        $this->authorize('delete', [Task::class, $project, $board, $task]);
+
         $task->delete();
         // unassign tags, assginees...
         return $this->jsonResponse($task);
@@ -100,6 +105,8 @@ class TasksController extends ApiController
 
     public function forceDelete(Project $project, Board $board, Task $task): JsonResponse
     {
+        $this->authorize('forceDelete', [Task::class, $project, $board, $task]);
+
         $task->forceDelete();
         // unassign tags, assginees...
         return $this->jsonResponse($task);
